backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image processing
pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
opencv-python>=4.8.0

# Model-specific libraries
# SAM2
segment-anything-2==0.1.0
torch==2.1.2
torchvision==0.16.2

# YOLOv8
ultralytics==8.1.0

# U-Net (using PyTorch)
# torch and torchvision already included above

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.92.2
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
